RAYSA-1 move check_if_bot_exists to util.py
