+++ Flask Web App/flasky/app.py
@@ -1,26 +1,9 @@
 from flask import Flask, render_template
 from flask import request
-import json 
+from util import  check_if_bot_exists 
 from flask_sqlalchemy import SQLAlchemy
 app = Flask(__name__)
 db = SQLAlchemy(app)
-
-def check_if_bot_exists(name):
-    file = open('bot_data.json')
-    data = json.load(file) 
-
-    # bots list 
-    bots_list = data["bots"]
-    bots_list_names = []
-    for bot in bots_list:
-        bots_list_names.append(bot["bot_name"])
-
-    #print(bots_list_names)
-
-    if name in bots_list_names: 
-        return True
-    else: 
-        return False
 
 @app.errorhandler(404)
 def page_not_found(content_name):
